+++ Cargo.toml
@@ -25,4 +25,4 @@ dirs = "5.0.1"
 chrono = "0.4.39"
 crossterm = { version = "0.28.1", features = ["event-stream"] }
 futures = "0.3.31"
-ratatui = "0.29.0"
+ratatui = { version = "0.29.0", features = ["all-widgets"] }
