+++ src/handler.rs
@@ -14,12 +14,8 @@ pub fn handle_key_events(key_event: KeyEvent, app: &mut App) -> AppResult<()> {
                 app.quit();
             }
         }
-        // Counter handlers
-        KeyCode::Right => {
-            app.increment_counter();
-        }
-        KeyCode::Left => {
-            app.decrement_counter();
+        KeyCode::Char('l') | KeyCode::Char('L') => {
+            app.file_list_widget.list_files();
         }
         // Other handlers you could add here.
         _ => {}
